fix(cli): resolve route generation import path mismatch for folder-based agents (#789)
This fix addresses the import path mismatch between registry.ts and routes.ts
generation for folder-based agents.

## Problem
When using bare imports like `@agent/hello`, the routes.ts generator assumed
the entry point was `index.js`, but the template created `agent.ts` with a
barrel `index.ts`. This caused TypeScript resolution failures.

## Solution
1. Simplified the agent template to use a single `index.ts` file containing
   the actual agent code (no more barrel file pattern)
2. Updated registry-generator to intelligently detect directory vs file imports
   and resolve paths accordingly
3. Added comprehensive tests for both patterns (index.ts and agent.ts)

## Changes
- `packages/cli/src/cmd/dev/templates.ts`: Create only `index.ts` with agent code
- `packages/cli/src/cmd/build/vite/registry-generator.ts`: Smart path resolution
- `templates/default/src/agent/translate/`: Moved code to `index.ts`, deleted `agent.ts`
- Added 3 new tests for backwards compatibility

## Backwards Compatibility
- Existing projects using `agent.ts` with explicit imports (`@agent/hello/agent`)
  continue to work
- New projects use the simpler single-file `index.ts` pattern
